소셜로그인 기능 구현 완료. username, email, password only save

+++ back/src/main/java/com/springboot/be/repository/RefreshRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshRepository extends JpaRepository<RefreshEntity, Integer> {
+public interface RefreshRepository extends JpaRepository<RefreshEntity, Long> {
 
     Optional<RefreshEntity> findByUser(User user);
     void deleteByUser(User user);
